# Conflicts:
#	Conciliacion/App_Web/Sitio Conicliacion 07-07-2017/Conciliacion.RunTime/DatosSQL/InformeBancarioDatos.cs

+++ Conciliacion/App_Web/Sitio Conicliacion 07-07-2017/Conciliacion.RunTime/ReglasDeNegocio/InformeBancario.cs
@@ -10,7 +10,7 @@ namespace Conciliacion.RunTime.ReglasDeNegocio
 {
     public abstract class InformeBancario : EmisorMensajes
     {
-        public List<DetalleCaja> consultarCajas(Conexion _conexion, int Caja)
+        public List<DetalleCaja> consultarCajas(Conexion _conexion, short Caja)
         {
             List<DetalleCaja> ListaResultado;
             DetalleCaja objDatos = new DetalleCaja();
